BOL-109 refactor: GroupResponse 위치 변경
